Ensure we set the working directory earlier
