+++ airbyte-ci/connectors/pipelines/pipelines/cli/airbyte_ci.py
@@ -6,16 +6,20 @@
 
 from __future__ import annotations
 
+# Important: This import and function call must be the first import in this file
+# This is needed to ensure that the working directory is the root of the airbyte repo
+from pipelines.cli.ensure_repo_root import set_working_directory_to_root
+
+set_working_directory_to_root()
+
 import logging
 import multiprocessing
 import os
 import sys
-from pathlib import Path
 from typing import Optional
 
 import asyncclick as click
 import docker  # type: ignore
-import git
 from github import PullRequest
 from pipelines import main_logger
 from pipelines.cli.auto_update import __installed_version__, check_for_upgrade, pre_confirm_auto_update_flag
@@ -28,58 +32,6 @@ from pipelines.dagger.actions.connector.hooks import get_dagger_sdk_version
 from pipelines.helpers import github
 from pipelines.helpers.git import get_current_git_branch, get_current_git_revision
 from pipelines.helpers.utils import get_current_epoch_time
-
-
-def _validate_airbyte_repo(repo: git.Repo) -> bool:
-    """Check if any of the remotes are the airbyte repo."""
-    expected_repo_name = "airbytehq/airbyte"
-    for remote in repo.remotes:
-        if expected_repo_name in remote.url:
-            return True
-
-    warning_message = f"""
-    ⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️
-
-    It looks like you are not running this command from the airbyte repo ({expected_repo_name}).
-
-    If this command is run from outside the airbyte repo, it will not work properly.
-
-    Please run this command your local airbyte project.
-
-    ⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️
-    """
-
-    logging.warning(warning_message)
-
-    return False
-
-
-def get_airbyte_repo() -> git.Repo:
-    """Get the airbyte repo."""
-    repo = git.Repo(search_parent_directories=True)
-    _validate_airbyte_repo(repo)
-    return repo
-
-
-def get_airbyte_repo_path_with_fallback() -> Path:
-    """Get the path to the airbyte repo."""
-    try:
-        repo_path = get_airbyte_repo().working_tree_dir
-        if repo_path is not None:
-            return Path(str(get_airbyte_repo().working_tree_dir))
-    except git.exc.InvalidGitRepositoryError:
-        pass
-    logging.warning("Could not find the airbyte repo, falling back to the current working directory.")
-    path = Path.cwd()
-    logging.warning(f"Using {path} as the airbyte repo path.")
-    return path
-
-
-def set_working_directory_to_root() -> None:
-    """Set the working directory to the root of the airbyte repo."""
-    working_dir = get_airbyte_repo_path_with_fallback()
-    logging.info(f"Setting working directory to {working_dir}")
-    os.chdir(working_dir)
 
 
 def log_context_info(ctx: click.Context) -> None:
@@ -240,8 +192,6 @@ async def airbyte_ci(ctx: click.Context) -> None:  # noqa D103
     if not ctx.obj["is_local"]:
         log_context_info(ctx)
 
-
-set_working_directory_to_root()
 
 if __name__ == "__main__":
     airbyte_ci()
